% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carlo_functions.R
\name{get_is_params}
\alias{get_is_params}
\title{Extract Importance Sampling Parameters}
\usage{
get_is_params(mc_res)
}
\arguments{
\item{mc_res}{An object of class "mc_result"}
}
\value{
A list of importance sampling parameters used for each dimension
}
\description{
Extract Importance Sampling Parameters
}
