% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carlo_functions.R
\name{mc_plot}
\alias{mc_plot}
\title{Plot precomputed Monte Carlo estimates}
\usage{
mc_plot(..., step = 1, true_value = NULL)
}
\arguments{
\item{...}{Named mc_result objects}

\item{step}{Number of samples per estimate (for x-axis scaling)}

\item{true_value}{Optional, true integral value}
}
\value{
ggplot of MC convergence
}
\description{
Plot precomputed Monte Carlo estimates
}
