% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carlo_functions.R
\name{preview_is_params}
\alias{preview_is_params}
\title{Preview Optimized Importance Sampling Parameters}
\usage{
preview_is_params(
  f,
  lower,
  upper,
  dim = NULL,
  distribution = c("normal", "beta", "exponential", "mixture_normal"),
  n_pilot = 1000
)
}
\arguments{
\item{f}{Function or expression}

\item{lower}{Numeric vector; lower bounds}

\item{upper}{Numeric vector; upper bounds}

\item{dim}{Integer; dimension (default inferred)}

\item{distribution}{Character; one of "normal", "beta", "exponential", "mixture_normal"}

\item{n_pilot}{Integer; number of pilot samples}
}
\description{
Preview Optimized Importance Sampling Parameters
}
