% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carlo_functions.R
\name{print.mc_result}
\alias{print.mc_result}
\title{Print method for Monte Carlo results}
\usage{
\method{print}{mc_result}(x, ...)
}
\arguments{
\item{x}{An object of class "mc_result"}

\item{...}{Ignored}
}
\description{
Print method for Monte Carlo results
}
