% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carlo_functions.R
\name{mc_print}
\alias{mc_print}
\title{Print Only Final Monte Carlo Estimate}
\usage{
mc_print(x)
}
\arguments{
\item{x}{mc_result or numeric vector}
}
\description{
Print Only Final Monte Carlo Estimate
}
