% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carlo_functions.R
\name{montecarlo_integrate}
\alias{montecarlo_integrate}
\title{Wrapper for Monte Carlo integration (vector of cumulative estimates)}
\usage{
montecarlo_integrate(
  f,
  lower,
  upper,
  n_samples,
  Partition = FALSE,
  dim = 1,
  step = 100
)
}
\arguments{
\item{f}{Function to integrate}

\item{lower}{Lower bound(s)}

\item{upper}{Upper bound(s)}

\item{n_samples}{Number of samples}

\item{Partition}{Use stratified partitioning (logical, not yet used in this version)}

\item{dim}{Dimension of integration}

\item{step}{Interval at which to record cumulative estimates}
}
\value{
An object of class "mc_result" containing:
  - estimates: numeric vector of cumulative estimates
  - final: last Monte Carlo estimate
}
\description{
Wrapper for Monte Carlo integration (vector of cumulative estimates)
}
