% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Carlo_functions.R
\name{print.mc_result}
\alias{print.mc_result}
\title{Print Monte Carlo Result}
\usage{
\method{print}{mc_result}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"mc_result"}}

\item{...}{Ignored}
}
\description{
Print Monte Carlo Result
}
